+++ src/javascript/defaults/store/api/api.endpoints.js
@@ -1,0 +1,9 @@
+const API = 'https://graviton-api.herokuapp.com' //http://localhost:3000 ,https://graviton-api.herokuapp.com
+const Search = `${API}/search`
+const List = `${API}/list`
+
+export default {
+	API,
+	Search,
+	List
+}