+++ src/javascript/defaults/store/api/get.list.js
@@ -1,0 +1,17 @@
+import Endpoints from './api.endpoints.js'
+import axios from 'axios'
+
+function getList(){
+	return new Promise((resolve,reject)=>{
+		axios({
+			method:'get',
+			url:Endpoints.List
+		}).then(async function (response) {
+			resolve(response.data.list)
+		}).catch((err)=>{
+			console.log(err)
+		})
+	})
+}
+
+export default getList